KBFS open external
