+++ desktop/app/paths.js
@@ -38,16 +38,3 @@ export function keybaseBinPath () {
   if (bundlePath === null) return null
   return path.resolve(bundlePath, 'Contents', 'SharedSupport', 'bin', 'keybase')
 }
-
-// pathToURL takes path and converts to (file://) url.
-// See https://github.com/sindresorhus/file-url
-export function pathToURL (path: string) {
-  path = path.replace(/\\/g, '/')
-
-  // Windows drive letter must be prefixed with a slash
-  if (path[0] !== '/') {
-    path = '/' + path
-  }
-
-  return encodeURI('file://' + path).replace(/#/g, '%23')
-}
